refactor: break circular dep in authSlice
This was screwing so many things up and had my head scratching for a few
days...

Good riddance I found a way to fix this with the trick described here:
https://medium.com/visual-development/how-to-fix-nasty-circular-dependency-issues-once-and-for-all-in-javascript-typescript-a04c987cf0de
